+++ liboscar-demos/learnopengl/AdvancedLighting/LOGLNormalMappingTab.cpp
@@ -9,36 +9,6 @@ using namespace osc;
 
 namespace
 {
-    // matches the quad used in LearnOpenGL's normal mapping tutorial
-    Mesh generate_quad()
-    {
-        Mesh mesh;
-        mesh.set_vertices({
-            {-1.0f,  1.0f, 0.0f},
-            {-1.0f, -1.0f, 0.0f},
-            { 1.0f, -1.0f, 0.0f},
-            { 1.0f,  1.0f, 0.0f},
-        });
-        mesh.set_normals({
-            {0.0f, 0.0f, 1.0f},
-            {0.0f, 0.0f, 1.0f},
-            {0.0f, 0.0f, 1.0f},
-            {0.0f, 0.0f, 1.0f},
-        });
-        mesh.set_tex_coords({
-            {0.0f, 1.0f},
-            {0.0f, 0.0f},
-            {1.0f, 0.0f},
-            {1.0f, 1.0f},
-        });
-        mesh.set_indices({
-            0, 1, 2,
-            0, 2, 3,
-        });
-        mesh.recalculate_tangents();
-        return mesh;
-    }
-
     MouseCapturingCamera create_camera()
     {
         MouseCapturingCamera camera;
@@ -56,7 +26,8 @@ namespace
         );
         const Texture2D normal_map = load_texture2D_from_image(
             loader.open("oscar_demos/learnopengl/textures/brickwall_normal.jpg"),
-            ColorSpace::Linear
+            ColorSpace::Linear,
+            ImageLoadingFlag::TreatComponentsAsSpatialVectors
         );
 
         Material rv{Shader{
@@ -84,7 +55,9 @@ public:
 
     explicit Impl(LOGLNormalMappingTab& owner, Widget* parent) :
         TabPrivate{owner, parent, static_label()}
-    {}
+    {
+        quad_mesh_.recalculate_tangents();  // needed for normal mapping
+    }
 
     void on_mount()
     {
@@ -146,7 +119,7 @@ private:
     Material normal_mapping_material_ = create_normal_mapping_material(loader_);
     Material light_cube_material_ = create_light_cube_material(loader_);
     Mesh cube_mesh_ = BoxGeometry{};
-    Mesh quad_mesh_ = generate_quad();
+    Mesh quad_mesh_ = PlaneGeometry{{.width = 2.0f, .height = 2.0f}};
 
     // scene state
     MouseCapturingCamera camera_ = create_camera();