+++ liboscar-demos/learnopengl/GettingStarted/LOGLTexturingTab.cpp
@@ -30,8 +30,7 @@ namespace
         {
             Texture2D container = load_texture2D_from_image(
                 loader.open("oscar_demos/learnopengl/textures/container.jpg"),
-                ColorSpace::sRGB,
-                ImageLoadingFlag::FlipVertically
+                ColorSpace::sRGB
             );
             container.set_wrap_mode(TextureWrapMode::Clamp);
 
@@ -42,8 +41,7 @@ namespace
         {
             const Texture2D face = load_texture2D_from_image(
                 loader.open("oscar_demos/learnopengl/textures/awesomeface.png"),
-                ColorSpace::sRGB,
-                ImageLoadingFlag::FlipVertically
+                ColorSpace::sRGB
             );
 
             rv.set("uTexture2", face);