+++ liboscar-demos/learnopengl/PBR/LOGLPBRLightingTexturedTab.cpp
@@ -46,7 +46,8 @@ namespace
         );
         const Texture2D normal = load_texture2D_from_image(
             loader.open("oscar_demos/learnopengl/textures/pbr/rusted_iron/normal.jpg"),
-            ColorSpace::Linear
+            ColorSpace::Linear,
+            ImageLoadingFlag::TreatComponentsAsSpatialVectors
         );
         const Texture2D metallic = load_texture2D_from_image(
             loader.open("oscar_demos/learnopengl/textures/pbr/rusted_iron/metallic.jpg"),
@@ -82,7 +83,9 @@ public:
 
     explicit Impl(LOGLPBRLightingTexturedTab& owner, Widget* parent) :
         TabPrivate{owner, parent, static_label()}
-    {}
+    {
+        sphere_mesh_.recalculate_tangents();  // normal mapping
+    }
 
     void on_mount()
     {